app_real_estate/pyproject.toml
[tool.poetry]
name = "app-real-estate"
version = "0.1.0"
description = ""
authors = ["tesseract <tesseractmaks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

fastapi = "^0.105.0"
sqlalchemy = "^2.0.23"
hypercorn = "^0.15.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pillow = "^10.1.0"
fastapi-pagination = "^0.12.13"
fastapi-filter = "^1.0.0"
pydantic-settings = "^2.1.0"
pymongo = "^4.6.1"
sqlalchemy-utils = "^0.41.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
websockets = "^12.0"
loguru = "^0.7.2"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.1"}
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
asgi-lifespan = "==2.*"
async-factory-boy = "^1.0.1"
numpy = "^1.26.3"
pytest-httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
